requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.35.7        # Fixed version
python-dotenv==1.0.1
pydantic==2.9.2
